refactor: use UnitConverter from typescript-crypto SDK
